1.2

+++ functions/be/fw-options.php
@@ -29,23 +29,27 @@ array( "name" => "Start Tabs",
 	array( "name" => "General Settings",
 			"id" => "tab_menu_1",
 			"type" => "tab",
-			"icon" => "layout"),
+			"icon" => "layout",
+			"class" => ""),
 	
 	// Social Networking Tab
 	array( "name" => "Social Networking",
 			"id" => "tab_menu_2",
 			"type" => "tab",
-			"icon" => "layout"),
+			"icon" => "layout",
+			"class" => ""),
 	
 	// Other settings Tab
 	array( "name" => "Other Settings",
 			"type" => "tab",
-			"id" => "tab_menu_3"),
+			"id" => "tab_menu_3",
+			"class" => ""),
 	
 	// SEO Tab
 	array( "name" => "SEO",
 			"type" => "tab",
-			"id" => "tab_menu_4"),
+			"id" => "tab_menu_4",
+			"class" => ""),
 	
 
 // General setting Tab
@@ -305,39 +309,31 @@ function wplook_admin_css() {
 	</script>
 
 <?php }
-
 function wpl_add_admin() {
 	global $options; global $themename; global $shortname;
-	if ( $_GET['page'] == basename(__FILE__) ) {
-		if ( 'save' == $_REQUEST['action'] ) {
-		//if(isset($_GET["action"])) {
-		
-		foreach ($options as $value) {
-			update_option( $value['id'], $_REQUEST[ $value['id'] ] ); 
-		}
-		
-		foreach ($options as $value) {
-			if( isset( $_REQUEST[ $value['id'] ] ) ) { update_option( $value['id'], $_REQUEST[ $value['id'] ] ); 
-			} else { 
-				delete_option( $value['id'] ); 
+	if ( isset ( $_GET['page'] ) && ( $_GET['page'] == basename(__FILE__) ) ) {
+		if ( isset ($_REQUEST['action']) && ( 'save' == $_REQUEST['action'] ) ){
+			foreach ( $options as $value ) {
+				if ( array_key_exists('id', $value) ) {
+					if ( isset( $_REQUEST[ $value['id'] ] ) ) {
+						update_option( $value['id'], $_REQUEST[ $value['id'] ]  );
+					}
+					else {
+						delete_option( $value['id'] );
+					}
+				}
 			}
-		}
-		
-		header("Location: themes.php?page=fw-options.php&saved=true");
-		die;
-		
-		} else if( 'reset' == $_REQUEST['action'] ) {
-		foreach ($options as $value) {
-			delete_option( $value['id'] ); }
-			header("Location: themes.php?page=fw-options.php&reset=true");
-		die;
+			header("Location: themes.php?page=".basename(__FILE__)."&saved=true");
+		}else if ( isset ($_REQUEST['action']) && ( 'reset' == $_REQUEST['action'] ) ) {
+			foreach ($options as $value) {
+				if ( array_key_exists('id', $value) ) {
+					delete_option( $value['id'] );
+				}
+			}
+			header("Location: themes.php?page=".basename(__FILE__)."&reset=true");
 		}
 	}
-	
-
-	
 	add_theme_page($themename." Options", 'Wplook Panel', 'edit_theme_options', 'fw-options.php', 'wpl_admin', 'http://i.wplook.com/fw-icon.jpg', '1');
-
 }
 
 
@@ -448,7 +444,7 @@ function wpl_admin() {
 	
 	<?php break; case "tabs-close":?>	
 	</div>	
-
+	
 	<?php break; case "tab":?>	
 	<div class="tab<?php echo $value['class']; ?>" id="<?php echo $value['id']; ?>">
 	<div class="link"><?php echo $value['name']; ?></div>
